# internal/transport/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Yaver API
  version: 0.1.0
  description: |
    Public API for Yaver. Session-cookie auth for the dashboard; API keys for
    server-to-server ingest. This spec is the source of truth for generated
    clients (yaver-web) and partner SDKs.
servers:
  - url: /
paths:
  /v1/auth/otp/request:
    post:
      operationId: requestOtp
      summary: Request a login OTP for a phone number
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequest'
      responses:
        '200':
          description: Code sent (dev_code present only in dev)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpRequestResult'
        '400': { $ref: '#/components/responses/BadRequest' }

  /v1/auth/otp/verify:
    post:
      operationId: verifyOtp
      summary: Verify an OTP and start a session (sets an httpOnly cookie)
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerify'
      responses:
        '200':
          description: Verified; session cookie set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionUser'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current session
      tags: [auth]
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'

  /v1/me:
    get:
      operationId: getMe
      summary: The authenticated user
      tags: [auth]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/calls:
    get:
      operationId: listCalls
      summary: Recent calls for the authenticated org
      tags: [calls]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Calls, newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallList'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/calls/{id}:
    get:
      operationId: getCall
      summary: A single call (org-scoped)
      tags: [calls]
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Call detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallDetail'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/BadRequest' }

  /v1/chat/messages:
    post:
      operationId: sendChatMessage
      summary: Send a message and get the assistant's reply
      tags: [chat]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                conversation_id: { type: string, description: Omit to start a new conversation }
                text: { type: string }
      responses:
        '200':
          description: Reply
          content:
            application/json:
              schema:
                type: object
                required: [conversation_id, reply]
                properties:
                  conversation_id: { type: string }
                  reply: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/chat/settings:
    get:
      operationId: getChatSettings
      summary: Chat/widget settings
      tags: [chat]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSettings'
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      operationId: saveChatSettings
      summary: Update chat/widget settings
      tags: [chat]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatSettings'
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Status' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/chat/conversations:
    get:
      operationId: listConversations
      summary: The org's chat conversations
      tags: [chat]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Conversations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationList'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/chat/conversations/{id}:
    get:
      operationId: getConversation
      summary: A conversation's messages
      tags: [chat]
      security:
        - cookieAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageList'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/BadRequest' }

  /v1/campaigns:
    get:
      operationId: listCampaigns
      summary: The org's campaigns
      tags: [campaigns]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Campaigns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignList'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createCampaign
      summary: Create a draft campaign
      tags: [campaigns]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { type: object, required: [id], properties: { id: { type: string } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/campaigns/{id}/start:
    post:
      operationId: startCampaign
      summary: Dispatch calls to all callable customers
      tags: [campaigns]
      security:
        - cookieAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Dispatched
          content:
            application/json:
              schema: { type: object, required: [queued], properties: { queued: { type: integer } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/BadRequest' }

  /v1/customers:
    get:
      operationId: listCustomers
      summary: The org's customers
      tags: [customers]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerList'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/customers/{id}/dnd:
    post:
      operationId: setDnd
      summary: Put a customer on do-not-disturb (skip calls)
      tags: [customers]
      security:
        - cookieAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: DND set
          content:
            application/json:
              schema: { type: object, properties: { dnd: { type: boolean } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      operationId: clearDnd
      summary: Remove a customer from do-not-disturb
      tags: [customers]
      security:
        - cookieAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: DND cleared
          content:
            application/json:
              schema: { type: object, properties: { dnd: { type: boolean } } }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/flows:
    get:
      operationId: listFlows
      summary: The org's flows
      tags: [flows]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Flows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowList'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/flows/{id}:
    get:
      operationId: getFlow
      summary: A flow with its spec (for the builder)
      tags: [flows]
      security:
        - cookieAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Flow detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowDetail'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/BadRequest' }
    put:
      operationId: updateFlow
      summary: Replace a flow's spec
      tags: [flows]
      security:
        - cookieAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spec]
              properties:
                spec: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/analytics/summary:
    get:
      operationId: getSummary
      summary: Dashboard metrics rollup
      tags: [analytics]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Summary'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/dev/test-call:
    post:
      operationId: testCall
      summary: Simulate a call with a keypress (no telephony) — generates an outcome
      tags: [calls]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCallRequest'
      responses:
        '200':
          description: Simulated call outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCallResult'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/settings/org:
    put:
      operationId: renameOrg
      summary: Rename the org (store) shown in the dashboard
      tags: [settings]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '200':
          description: Renamed
          content:
            application/json:
              schema:
                type: object
                required: [name]
                properties:
                  name: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/settings/publishable-key:
    post:
      operationId: createPublishableKey
      summary: Mint a publishable (yvr_pk_) key for the embeddable widget
      tags: [settings]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Publishable key (shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResult'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/settings/api-keys:
    get:
      operationId: listApiKeys
      summary: List the org's API keys (metadata only)
      tags: [settings]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createApiKey
      summary: Create an API key for the org (returned once)
      tags: [settings]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: The new key (store it now — not retrievable later)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResult'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/settings/webhook:
    get:
      operationId: getWebhook
      summary: Current webhook config (no secret)
      tags: [settings]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Webhook config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookConfig'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: setWebhook
      summary: Set the webhook URL; returns a fresh signing secret once
      tags: [settings]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookSet'
      responses:
        '200':
          description: Saved; signing secret (store it now)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSecret'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/events:
    post:
      operationId: ingestEvent
      summary: Ingest a merchant store event (triggers a call for order_placed)
      tags: [ingest]
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventAccepted'
        '200':
          description: Duplicate (idempotent no-op)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: yaver_session
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Status:
      type: object
      required: [status]
      properties:
        status: { type: string }
    OtpRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string, example: '01712345678' }
    OtpRequestResult:
      type: object
      required: [status]
      properties:
        status: { type: string }
        dev_code:
          type: string
          description: Present only in dev (no SMS provider yet)
    OtpVerify:
      type: object
      required: [phone, code]
      properties:
        phone: { type: string }
        code: { type: string }
    SessionUser:
      type: object
      required: [user_id, phone]
      properties:
        user_id: { type: string }
        phone: { type: string }
    Me:
      type: object
      required: [user_id, phone, email, name, org]
      properties:
        user_id: { type: string }
        phone: { type: string }
        email: { type: string }
        name: { type: string }
        org:
          type: object
          required: [id, name]
          properties:
            id: { type: string }
            name: { type: string }
    Summary:
      type: object
      required: [total, confirmed, cancelled, today, confirmation_rate]
      properties:
        total: { type: integer }
        confirmed: { type: integer }
        cancelled: { type: integer }
        today: { type: integer }
        confirmation_rate: { type: integer, description: percent 0-100 }
    TestCallRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string }
        digit: { type: string, description: 'simulated keypress: 1|2|3' }
    TestCallResult:
      type: object
      required: [call_id, status, result]
      properties:
        call_id: { type: string }
        status: { type: string }
        result: { type: string }
    ApiKeyResult:
      type: object
      required: [api_key]
      properties:
        api_key: { type: string }
    ApiKeyInfo:
      type: object
      required: [prefix, name, kind, created_at]
      properties:
        prefix: { type: string }
        name: { type: string }
        kind: { type: string, enum: [secret, publishable] }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time, nullable: true }
    ApiKeyList:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items: { $ref: '#/components/schemas/ApiKeyInfo' }
    WebhookConfig:
      type: object
      required: [configured, events]
      properties:
        configured: { type: boolean }
        url: { type: string }
        events:
          type: array
          items: { type: string }
    WebhookSet:
      type: object
      required: [url]
      properties:
        url: { type: string }
        events:
          type: array
          items: { type: string }
    WebhookSecret:
      type: object
      required: [secret]
      properties:
        secret: { type: string }
    ChatSettings:
      type: object
      required: [instructions, widget_title, welcome, accent]
      properties:
        instructions: { type: string }
        widget_title: { type: string }
        welcome: { type: string }
        accent: { type: string }
    Conversation:
      type: object
      required: [id, status, last_message, message_count, updated_at]
      properties:
        id: { type: string }
        status: { type: string }
        last_message: { type: string }
        message_count: { type: integer }
        updated_at: { type: string, format: date-time }
    ConversationList:
      type: object
      required: [conversations]
      properties:
        conversations:
          type: array
          items: { $ref: '#/components/schemas/Conversation' }
    ChatMessage:
      type: object
      required: [role, content, created_at]
      properties:
        role: { type: string, enum: [user, assistant, system] }
        content: { type: string }
        created_at: { type: string, format: date-time }
    MessageList:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items: { $ref: '#/components/schemas/ChatMessage' }
    Campaign:
      type: object
      required: [id, name, status, target_count, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string }
        target_count: { type: integer }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
    CampaignList:
      type: object
      required: [campaigns]
      properties:
        campaigns:
          type: array
          items: { $ref: '#/components/schemas/Campaign' }
    Customer:
      type: object
      required: [id, phone, name, dnd, created_at]
      properties:
        id: { type: string }
        phone: { type: string }
        name: { type: string }
        dnd: { type: boolean }
        created_at: { type: string, format: date-time }
    CustomerList:
      type: object
      required: [customers]
      properties:
        customers:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
    FlowSummary:
      type: object
      required: [id, name, version, channel, type, active]
      properties:
        id: { type: string }
        name: { type: string }
        version: { type: integer }
        channel: { type: string }
        type: { type: string }
        active: { type: boolean }
    FlowList:
      type: object
      required: [flows]
      properties:
        flows:
          type: array
          items: { $ref: '#/components/schemas/FlowSummary' }
    FlowDetail:
      type: object
      required: [id, name, version, channel, type, locale, spec]
      properties:
        id: { type: string }
        name: { type: string }
        version: { type: integer }
        channel: { type: string }
        type: { type: string }
        locale: { type: string }
        spec: { type: object, additionalProperties: true }
    Call:
      type: object
      required: [id, direction, status, result, created_at]
      properties:
        id: { type: string }
        direction: { type: string, enum: [inbound, outbound] }
        status:
          type: string
          enum: [queued, ringing, in_progress, completed, failed, no_answer]
        result: { type: string }
        created_at: { type: string, format: date-time }
    CallList:
      type: object
      required: [calls]
      properties:
        calls:
          type: array
          items: { $ref: '#/components/schemas/Call' }
    CallDetail:
      type: object
      required: [id, direction, status, result, flow_id, provider_call_id, created_at]
      properties:
        id: { type: string }
        direction: { type: string }
        status: { type: string }
        result: { type: string }
        flow_id: { type: string }
        provider_call_id: { type: string }
        created_at: { type: string, format: date-time }
    EventRequest:
      type: object
      required: [event_type, event_id]
      properties:
        event_type:
          type: string
          enum: [order_placed, order_cancelled, abandoned_cart]
        event_id: { type: string, description: Merchant-side unique id (idempotency) }
        customer:
          type: object
          properties:
            phone: { type: string }
    EventAccepted:
      type: object
      required: [status, event_id]
      properties:
        status: { type: string }
        event_id: { type: string }